Fixing errors wih db:migrate

+++ db/migrate/20151119124111_drop_collection_rating_table.rb
@@ -1,5 +1,5 @@
 class DropCollectionRatingTable < ActiveRecord::Migration
   def change
-    drop_table :collection_ratings
+    drop_table :collection_ratings if ActiveRecord::Base.connection.table_exists? 'collection_ratings'
   end
 end
